+++ docs/specs/261-market-hours-alert-filter.md
@@ -1,0 +1,49 @@
+# 거래시간 기반 변동 알림 필터링
+
+## 목적
+
+한국장 거래시간이 아닐 때(평일 15:30 이후, 주말 등) Yahoo Finance 캐시 갱신으로 인해 발생하는
+허위 급등락 알림을 차단.
+
+## 문제
+
+- `checkPriceAlerts`가 시세 갱신 후 종목 시장 구분 없이 급등락 체크
+- 시세 갱신 cron은 장외에도 1시간 간격으로 실행
+- 한국 종목의 경우 야간에 Yahoo Finance 캐시 업데이트 시 종가 추정치 변동이 발생
+  → 실제 거래가 아닌데도 "급등/급락" 판정해서 알림 발송
+
+## 요구사항
+
+- [ ] 한국 종목(market=KR)은 한국장 거래시간(평일 09:00~15:30 KST)에만 급등락 알림
+- [ ] 미국 종목(market=US)은 미국장 거래시간에만 급등락 알림 (기존 DST 체크 재사용)
+- [ ] 환율(USDKRW=X)은 24시간 거래이므로 제한 없음
+- [ ] 목표가/손절가/매수구간 도달 알림은 제한 없이 유지 (사용자 설정 기반이므로 언제든 유의미)
+
+## 기술 설계
+
+### 변경 파일
+
+1. **`src/lib/market-hours.ts`** (신규) — 시장별 거래시간 판별 유틸 (cron.ts에서 이동/추출)
+   - `isKRMarketOpen()`, `isUSMarketOpen()` export
+   - cron.ts에서는 이 유틸 사용
+
+2. **`src/bot/notifications/price-alert.ts`** — 급등락 체크 시 market 기반 거래시간 가드
+   - 보유 종목 조회 시 `market` 필드 포함
+   - 급등락 알림 조건: `market === 'KR'` → `isKRMarketOpen()` 필요, `market === 'US'` → `isUSMarketOpen()` 필요
+   - 목표가/매수구간 알림은 기존대로 유지
+
+3. **`src/lib/cron.ts`** — market-hours.ts에서 함수 import로 변경
+
+## 테스트 계획
+
+- [ ] 한국장 거래시간 중 → 한국 종목 급등락 알림 발송
+- [ ] 한국장 종료 후 → 한국 종목 급등락 알림 스킵
+- [ ] 미국장 시간 → 미국 종목 알림 발송
+- [ ] 환율 변동 알림 → 시간 무관 발송
+- [ ] 목표가 도달 알림 → 시간 무관 발송
+- [ ] lint + typecheck + build 통과
+
+## 제외 사항
+
+- 개별 종목 알림 ON/OFF 토글 (별도 이슈)
+- 실시간/추정가 구분 (기준 모호, 별도 이슈)
