+++ src/lib/cron.ts
@@ -5,6 +5,7 @@ import { syncKrxStocks } from './krx-stocks'
 import { prisma } from './prisma'
 import { checkPriceAlerts } from '@/bot/notifications/price-alert'
 import { checkTASignals } from '@/bot/notifications/ta-signal-alert'
+import { isKRMarketOpen, isUSMarketOpen } from './market-hours'
 import { calculateNextRunAt, type Frequency } from './recurring-utils'
 import { sendToWhooing } from './whooing-webhook'
 
@@ -48,51 +49,11 @@ function createCronGuard(label: string, timeoutMs: number) {
 }
 
 /** 현재 시각을 KST 기준으로 반환 (시, 분, 요일) */
-function getKSTTime(): { h: number; m: number; day: number } {
+/** KST 기준 분 단위 시간 (cron 내부 전용) */
+function getKSTMinutes(): number {
   const now = new Date()
   const kst = new Date(now.toLocaleString('en-US', { timeZone: 'Asia/Seoul' }))
-  return {
-    h: kst.getHours(),
-    m: kst.getMinutes(),
-    day: kst.getDay(), // 0=Sun, 6=Sat
-  }
-}
-
-/** 한국 주식시장 개장 여부 (09:00~15:30 KST, 평일) */
-function isKRMarketOpen(): boolean {
-  const { h, m, day } = getKSTTime()
-  if (day === 0 || day === 6) return false
-  const minutes = h * 60 + m
-  return minutes >= 540 && minutes <= 930 // 09:00 ~ 15:30
-}
-
-/** 현재 미국 서머타임(EDT) 여부를 America/New_York 시간대로 판별 */
-function isUSDST(): boolean {
-  const now = new Date()
-  const et = new Date(now.toLocaleString('en-US', { timeZone: 'America/New_York' }))
-  const utc = new Date(now.toLocaleString('en-US', { timeZone: 'UTC' }))
-  const diffHours = (utc.getTime() - et.getTime()) / (1000 * 60 * 60)
-  return Math.round(diffHours) === 4 // EDT=UTC-4, EST=UTC-5
-}
-
-/**
- * 미국 주식시장 개장 여부 (평일→익일, KST 기준)
- * - 서머타임(EDT): 22:30~05:00 KST
- * - 비서머타임(EST): 23:30~06:00 KST
- */
-function isUSMarketOpen(): boolean {
-  const { h, m, day } = getKSTTime()
-  const minutes = h * 60 + m
-  const dst = isUSDST()
-  const openMin = dst ? 1350 : 1410   // 22:30 or 23:30
-  const closeMin = dst ? 300 : 360    // 05:00 or 06:00
-
-  // 저녁 세션: openMin~23:59 (월~금)
-  if (day >= 1 && day <= 5 && minutes >= openMin) return true
-  // 새벽 세션: 00:00~closeMin (화~토 = 월~금 야간)
-  if (day >= 2 && day <= 6 && minutes <= closeMin) return true
-
-  return false
+  return kst.getHours() * 60 + kst.getMinutes()
 }
 
 /** TA 체크 주기 제어: DB 설정(ta_check_interval_min) 기반 */
@@ -121,9 +82,9 @@ export function schedulePriceUpdates(): void {
   cron.schedule(
     '*/10 * * * *',
     async () => {
-      const { m } = getKSTTime()
+      const kstMinutes = getKSTMinutes()
       const isMarketHours = isKRMarketOpen() || isUSMarketOpen()
-      const isTopOfHour = m < 10 // 매시 0~9분 구간
+      const isTopOfHour = kstMinutes % 60 < 10 // 매시 0~9분 구간
 
       if (isMarketHours || isTopOfHour) {
         try {
